Cors añadido + api acabada

+++ src/controllers/adminCtrl.ts
@@ -1,19 +1,6 @@
 import { Request, Response } from "express";
 import { Evento } from "../models/Evento.js";
-
-/* export const getEventos = async (req: Request, res: Response) => {
-    try {
-        const eventos = await Evento.fetchAll();
-        res.json({ eventos });
-    } catch (error: any) {
-        res.status(500).json({ error: error.message });
-    }
-}; */
-
-/* export const getAddEvento = (req: Request, res: Response) => {
-    console.log("Devolvemos el formulario para agregar eventos");
-    res.render('admin/edit-evento', { pageTitle: "Formulario de Evento", path: "/admin/add-evento", editing: false });
-}; */
+import { Order } from "../models/Orders.js";
 
 export const createEvento = async (req: Request, res: Response) => {
     const { nombreEvento, fechaEvento, lugar, descripcion, precio } = req.body;
@@ -26,26 +13,6 @@ export const createEvento = async (req: Request, res: Response) => {
         res.status(500).json({ error: error.message });
     }
 };
-
-/* export const getEditEvento = async (req: Request, res: Response) => {
-    console.log("getEditEvento: Devolvemos el formulario para editar eventos");
-    const editMode = req.query.edit === 'true';
-    if (!editMode) {
-        return res.redirect('/eventos');
-    }
-    const eventoId = req.params.eventoId;
-    const evento = await Evento.findById(eventoId);
-    if (evento) {
-        res.render('admin/edit-evento', {
-            pageTitle: "Formulario de Edición de Evento",
-            path: "/admin/edit-evento",
-            editing: editMode,
-            evento: evento
-        });
-    } else {
-        res.redirect('/eventos');
-    }
-}; */
 
 export const updateEvento = async (req: Request, res: Response) => {
     const eventoId = req.params.eventoId;
@@ -72,9 +39,22 @@ export const updateEvento = async (req: Request, res: Response) => {
 export const deleteEvento = async (req: Request, res: Response) => {
     const eventoId = req.params.eventoId;
     try {
+        const evento = await Evento.findById(eventoId);
+        if (!evento) {
+            return res.status(404).json({ message: 'Evento no encontrado' });
+        }
         await Evento.deleteById(eventoId);
         res.status(200).json({ message: 'Evento eliminado exitosamente' });
     } catch (error: any) {
         res.status(500).json({ error: error.message });
     }
 };
+
+export const getAllOrders = async (req: Request, res: Response) => {
+    try {
+        const orders = await Order.fetchAll();
+        res.json({ orders });
+    } catch (error: any) {
+        res.status(500).json({ error: error.message });
+    }
+}
